requirements.txt
streamlit
scipy
numpy
Pillow
pandas
matplotlib
svgpathtools